update to cope version that has default implementations for interface methods, bugfix
was:
integrate cope support for long pks

git-svn-id: https://svn.exedio.com/cope/console/branches/changeable_PK_database_type@29915 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ src/com/exedio/cope/console/TestCop.java
@@ -208,7 +208,7 @@ abstract class TestCop<I> extends ConsoleCop<HashMap<String, TestCop.Info>>
 	{
 		int elapsed = 0;
 		Info date = null;
-		int failures = 0;
+		long failures = 0;
 		boolean isError = false;
 
 		for(final Info info : infos.values())
